resource name comming from one file
